Seprate playback components from AudioService class.
